Move SignalComponentAutocomplete.js to components folder and update imports
